Multiple changes towards a better handling of hardware devices: - changed adc and temp sensor instantiations to shared_ptr of more generic class type DeviceFunction - added search range for static findI2cDeviceType function, which searches for specific i2c devices and returns their addresses when found - added identify method to eeprom device - temperature sensor is now instatiated according to the specific device that was actually found on the bus, i.e. either an LM75 or MIC184 - temp sensor, adc and eeprom are now not bound to a fixed i2c address but scanned for on the i2c bus

+++ daemon/src/hardware/i2c/ads1115.cpp
@@ -222,7 +222,7 @@ ADS1115::Sample ADS1115::getSample( unsigned int channel )
     return sample;
 }
 
-bool ADS1115::triggerConversion( uint8_t channel )
+bool ADS1115::triggerConversion( unsigned int channel )
 {
 	// triggering a conversion makes only sense in single shot mode
 	if ( fConvMode == CONV_MODE::SINGLE /* && !fTriggered */ ) {
